+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.8.0] - TODO
+### Fixes
+- Fixed macro hygiene with `Err`
+- Fixed some bugs with formatting
+
+### Changes
+- `StackedError` is now special cased to attempt to downcast whenever used as an argument on another `StackedError`, and the two chains are combined to prevent nesting
+- `Default` for `StackedError` uses `Error::empty()` (which does not capture a location and does not allocate) instead of `Error::new()` now
+
 ## [0.7.2] - 2026-08-11
 ### Fixes
 - Fixed future breakage from the macros on nightly (https://github.com/rust-lang/rust/issues/79813)